requirements.txt
pandas
requests
beautifulsoup4
python-whois
streamlit
numpy
matplotlib
seaborn
xgboost
scikit-learn
tensorflow
keras
